move score parameter types to their own file
